Correções na qtde de pedido, ainda com dificuldades para atualizar.

+++ src/screens/Pedidos/Pedido.js
@@ -43,7 +43,7 @@ const Pedido = props => {
                index = pedido.indexOf(item);
             }
         })
-        pedido[index].qtde =  parseInt(pedido[index].qtde) + 1 ;
+        pedido[index].qtde =  parseFloat(pedido[index].qtde) + 1 ;
         let valorFinal = parseFloat(pedido[index].vlr_unidade) * pedido[index].qtde;
         pedido[index].vlr_vendido = valorFinal;
 
@@ -59,7 +59,7 @@ const Pedido = props => {
           }
       })
 
-      pedido[index].qtde > 1 ? pedido[index].qtde =  parseInt(pedido[index].qtde) - 1 : "";
+      pedido[index].qtde > 1 ? pedido[index].qtde =  parseFloat(pedido[index].qtde) - 1 : "";
       let valorFinal = parseFloat(pedido[index].vlr_unidade) * pedido[index].qtde;
       pedido[index].vlr_vendido = valorFinal;
       props.setPedido(pedido)
